修正CPU3参数单位倍率与手输值存储（CPU2 V1.12.1.2 / CPU3 V1.11.1.2）
版本：
- CPU2：保持 V1.12.1.2。
- CPU3：V1.11.1.1 -> V1.11.1.2。

协议版本/兼容性：
- 该历史提交未单列协议或兼容性说明；本次说明改写不改变该提交的文件树和程序行为。

本次修改：
- 背景：CPU3参数菜单中部分单位/倍率口径与CPU2实际算法不一致，密度手输值显示倍率和本地存储宽度也存在截断风险。
- 主要改动：补齐电机限速、液位阈值、水位电容、探底角度等参数单位；将液位找液/滞后阈值改为整数Hz显示；将密度手输值修正为kg/m3 x10口径；将CPU3本地手输值字段扩展为int32_t并提升CPU3本地参数版本到0x0003；CPU3固件版本升至V1.11.1.2。
- 文档：同步更新CHANGELOG、系统参数出厂默认值、CPU3参数单位与范围清单、版本改动与测试方案及索引。

验证：
- py LTD_DISPLAY_CPU3\font_check.py通过；cmake --build build\LTD_DISPLAY_CPU3通过并生成LTD_DISPLAY_CPU3_V1.11.1.2.hex；git diff --cached --check通过；py tools\check_version_bumped.py通过。
- 影响范围：不改变DEVICE_PROTOCOL_VERSION，不改变CPU2 DeviceParameters结构和CPU2参数存储版本；CPU3本地显示/通信参数结构变更，升级后会按默认值重建CPU3本地参数。

+++ LTD_DISPLAY_CPU3/Application/system_param/cpu3_comm_display_params.h
@@ -28,11 +28,12 @@ typedef struct
     uint8_t screen_source_t;
 
     /* ---------- 手工输入 ---------- */
-    uint8_t screen_input_oil;
-    uint8_t screen_input_water;
-    uint8_t screen_input_d;
+    /* 手输值按菜单原始倍率保存，不能用 uint8_t 截断。 */
+    int32_t screen_input_oil;
+    int32_t screen_input_water;
+    int32_t screen_input_d;
     uint8_t screen_input_d_switch;
-    uint8_t screen_input_t;
+    int32_t screen_input_t;
 
     /* ---------- 显示类参数 ---------- */
     uint8_t  screen_decimal;
